Fix MISRA issues
:Release Notes:
Fix MISRA issues

:Detailed Notes:
Address MSIRA issues for the below
mentioned CIDs

[CID-6527191] MISRA C++-2008 Expressions
[CID-6528034] MISRA C++-2008 Expressions
[CID-6528301] MISRA C++-2008 Declarators
[CID-6528480] MISRA C++-2008 Expressions
[CID-6528763] MISRA C++-2008 Expressions
[CID-6529104] MISRA C++-2008 Expressions
[CID-6529214] MISRA C++-2008 Expressions
[CID-6529215] MISRA C++-2008 Declarators
[CID-6529550] MISRA C++-2008 Declarators
[CID-6530047] MISRA C++-2008 Expressions
[CID-6530174] MISRA C++-2008 Expressions
[CID-6530289] MISRA C++-2008 Expressions
[CID-6531564] MISRA C++-2008 Expressions
[CID-6532505] MISRA C++-2008 Expressions
[CID-6532820] MISRA C++-2008 Declarators
[CID-6533244] MISRA C++-2008 Declarators
[CID-6533675] MISRA C++-2008 Expressions
[CID-6534086] MISRA C++-2008 Expressions
[CID-6535381] MISRA C++-2008 Expressions
[CID-6535383] MISRA C++-2008 Expressions
[CID-6535946] MISRA C++-2008 Declarators
[CID-6536268] MISRA C++-2008 Expressions
[CID-6536628] MISRA C++-2008 Declarators
[CID-6537644] MISRA C++-2008 Expressions
[CID-6537752] MISRA C++-2008 Expressions
[CID-6539557] MISRA C++-2008 Expressions
[CID-6539940] MISRA C++-2008 Expressions
[CID-6540182] MISRA C++-2008 Expressions
[CID-6540353] MISRA C++-2008 Expressions
[CID-6540890] MISRA C++-2008 Expressions
[CID-6540981] MISRA C++-2008 Expressions
[CID-2440429] MISRA C-2012 Initializers
[CID-2446158] MISRA C-2012 Initializers
[CID-4577958] MISRA C-2012 Initializers
[CID-8985602] MISRA C-2012 Initializers
[CID-8985603] MISRA C-2012 Initializers
[CID-8985604] MISRA C-2012 Initializers
[CID-8985605] MISRA C-2012 Initializers
[CID-6528140] MISRA C++-2008 Basic Concepts
[CID-6536566] MISRA C++-2008 Basic Concepts
[CID-6538210] MISRA C++-2008 Basic Concepts
[CID-6538326] MISRA C++-2008 Basic Concepts
[CID-6531517] MISRA C++-2008 Basic Concepts

:Testing Performed:
Build successful

:QA Notes:
NA

:Issues Addressed:
[PLAT-138084] Fix MISRA Issues - libpbnjson

+++ src/pbnjson_validate/pbnjson_validate.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2018 LG Electronics, Inc.
+// Copyright (c) 2014-2021 LG Electronics, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ namespace {
 
 const char *Basename(const char *path)
 {
-	const char *res = strrchr(path, '/');
+	const char *res = strrchr(path, (int)('/'));
 	return res ? res + 1 : path;
 }
